Create MusicLibrary
The MusicLibrary is a place for all loaded
music. It can be given a SongId -> Song map
or it can load all the music under a given
directory.

The music dir has been moved to /opt/domiugo/espressivo/music
as /opt/domiugo is the new home for all domiugo projects
and espressivo is the home for this project